ArduPlane: Fix bugs in flight mode select logic.
Ensure that if in the VTOL or FW camera control mode  a transition is always to the corresponding default FW or VTOL camera mode.
Ensure that if in a non-stick controlled FW mode, the first transition is into CRUISE.
Ensure that if in a non stick controlled VTOL or FW mode, the first transition is always into the corresponding default VTOL or FW trajectory control mode.