AP_Landing: Fix premature transiton after WP overshoot
Selection of RTL from downwind of the recovery point at a height exeding the Q_RTL height generates a WP overshoot event and "Recovery - Overshot WP - Using Loiter" message.  This casued the  vehicle to go into a loiter recovery to enable the vehicle to spiral down to the transition height to save power.

This was causing  _target_heading_deg to be used at a incorrectly intialised value that was 90 degrees out.