AP_TECS: Remove hardcoded time step assumptions
TECS has a combination of a hardcoded 0.1sec time step and a measured time step.
This change replaces all instances where a 0.1 sec time step is assumed with use of the measured value.
The starting vlaue before measurement is set to be consistent with the default atomic update rate for planes of 50Hz.