ArduPlane: Rework loiter tangent exit calculation
Uses EKF ground course estimate instread of GPS data which is only updated at 5Hz and can be affected by glitches.
Prevents premature exit when next waypoint is close by reworking the test that detects if the next waypoint is inside the turn and therefore no tangent exit is possible.
Notifies the console if premature exit has occurred.
Exits with time allowance for roll to level.