AP_TECS: Use separate method for controlling height without airspeed sensor
Without airspeed sensing we cannot use total energy and energy balance to calculate the pitch angle demand so the existing method which used the demanded rate of change of energy balance to set pitch angle did not produce the correct time constant when operating without airspeed sensing.

This adds a dedicated function which is used to calculate the demanded pitch angle when the total energy control with airspeed sensing is not available.