ArduPlane: Fix bug causing approach from downwind
SITL testing with a wind from the NE, p1=180 and p2=90 caused the vehicle to approach from the west.
This was caused by the way that the into wind approach angle was being checked against sector limits.
This change calculates the delta between the into wind yaw and the sector yaw and compares that to the tolerance value.