ArduPlane: Rework use of GPS velocity and airspeed for is flying check
Limit use of airspeed estimate to when there is an airspeed sensor being used to form the estimate. Otherwise the estimate is unreliable in changing wind conditions or at launch when wind is unknown.
Use GPS reported accuracy to determine the velocity threshold used to check when stopping flight
Use a higher threshold based on the min ground speed parameter to check when starting flight
Make consequence of losing GPS lock clearer.

TODO - add check of IMU data which needs a modification to the inertial sensor library to capture required statistics.