AP_NavEKF3: Rework use of on ground not moving yaw
Use consistent value for minimum allowed yaw observation variance
If GPS yaw is lost when on ground and not moving, fuse a stored static yaw reference every 140 msec as a synthetic observation to prevent yaw drift and improve learning of yaw gyro bias.
Move the calculation and saving of the static yaw reference into a function is which called immediately after the update of the quaternion states in the INS prediction when storage request is received.
Always refresh the stored static yaw reference if a yaw reset is performed.
Move generation of a the yaw storage request into the same function that determines if on ground and not moving to eliminate possibility of a race condition causing use of the static yaw reference before it has been stored.